requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
gunicorn
psycopg2
gdown
click
flask
jinja2
markupsafe
itsdangerous
flask_cors
werkzeug
torch==1.9.0+cpu
torchvision==0.10.0+cpu
torchaudio==0.9.0
Pillow
opencv-python-headless==4.2.0.32
scikit-image
python-Levenshtein
pymongo[srv]
craft-text-detector
vietocr==0.3.5
rapidfuzz
pyyaml
sklearn
nltk
tensorflow-cpu
